For verbose output, beware of "onto" within a word
If a RwGPS cue contains a string such as "Turn right onto Promontory
Road", the debug log would say "y Road", naively chopping off everything
to the left of 5 characters past the second occurrence of "onto".  The
easiest fix is to get rid of the conditional string manipulation.

Also, use "km" as the standard SI abbreviation for kilometers.